+++ ror/cmd/ror/delete_cli.go
