+++ ror/cmd/ror/start_cli.go
@@ -1,9 +1,26 @@
-//go:build !linux
-
 package main
 
-import "fmt"
+import (
+	"context"
+	"fmt"
 
-func startContainer(id string) error {
-	return fmt.Errorf("cannot start container on non-linux OS")
+	"github.com/urfave/cli/v3"
+)
+
+func newStartCmd() *cli.Command {
+	return &cli.Command{
+		Name:      "start",
+		Usage:     "Start a created container",
+		ArgsUsage: "<id>",
+		Action: func(c context.Context, cmd *cli.Command) error {
+			if cmd.Args().Len() < 1 {
+				return cli.Exit("container id required", 1)
+			}
+
+			fmt.Printf("[Starting Container::] %s\n", cmd.Args().First())
+			id := cmd.Args().First()
+
+			return startContainer(id)
+		},
+	}
 }
