+++ ror/internal/runner/create.go
@@ -1,7 +1,6 @@
-package main
+package runner
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"log"
@@ -9,39 +8,9 @@ import (
 	"path/filepath"
 
 	"github.com/opencontainers/runtime-spec/specs-go"
-	"github.com/urfave/cli/v3"
 )
 
-func newCreateCmd() *cli.Command {
-	var cfg ContainerConfig
-
-	return &cli.Command{
-		Name:      "create",
-		Usage:     "Create an OCI container",
-		ArgsUsage: "<id>",
-		Flags: []cli.Flag{
-			&cli.StringFlag{
-				Name:        "bundle",
-				Aliases:     []string{"b"},
-				Value:       ".",
-				Usage:       "Path to the OCI bundle",
-				Destination: &cfg.Bundle,
-			},
-			&cli.StringFlag{
-				Name:        "pid-file",
-				Usage:       "write child pid to this file (optional)",
-				Destination: &cfg.PIDFile,
-			},
-		},
-		Action: func(c context.Context, cmd *cli.Command) error {
-			cfg.ID = cmd.Args().First()
-			cfg.BasePath = defaultBasePath
-			return createContainer(cfg)
-		},
-	}
-}
-
-func createContainer(cfg ContainerConfig) error {
+func (r *Runner) CreateContainer(cfg ContainerConfig) error {
 	if cfg.ID == "" {
 		return fmt.Errorf("container id required")
 	}
