+++ ror/internal/runner/init_linux.go
@@ -1,6 +1,6 @@
 //go:build linux
 
-package main
+package runner
 
 import (
 	"encoding/json"
