+++ ror/internal/runner/init_other.go
@@ -1,0 +1,10 @@
+//go:build !linux
+
+package runner
+
+import "fmt"
+
+// This is the dummy function for non-Linux systems.
+func initContainer(id string) error {
+	return fmt.Errorf("init command is for internal use on linux only")
+}
