+++ ror/internal/runner/runner.go
@@ -1,7 +1,9 @@
 package runner
 
 // Runner will orchestrate namespaces/cgroups later
-type Runner struct{}
+type Runner struct {
+	BasePath string
+}
 
 func New() *Runner {
 	return &Runner{}
